refactor: 担当者別ガントチャートを汎用コンポーネントへ抽出
- WorkloadGanttChart: fetcher駆動の汎用チャート(週/月ナビ・詳細Sheet・警告表を移設)
- gantt-header/row/cell を src/components/assignee-gantt/ へ移動しprojectName表示に対応
- AssigneeGanttChart は公開props不変の薄いコンテナ化(3つの埋め込み箇所は無修正)
- workload-data.ts に AssigneeWarningData/AssigneeGanttResponse を移設、共有マッパー toWorkloadData を新設
